requirements.txt
streamlit
pytesseract
Pillow
PyMuPDF
langchain
chromadb
python-dotenv
tavily-python
langchain-google-genai
langchain-community